sql: create `sql/clustermode` package
This commit is to move the cluster mode and related info to a seperate new
package, `sql/clustermode`.

Release note: none
