+++ pkg/BUILD.bazel
@@ -1257,6 +1257,7 @@ GO_TARGETS = [
     "//pkg/sql/catalog/typedesc:typedesc_test",
     "//pkg/sql/catalog:catalog",
     "//pkg/sql/catalog:catalog_test",
+    "//pkg/sql/clustermode:clustermode",
     "//pkg/sql/clusterunique:clusterunique",
     "//pkg/sql/colcontainer:colcontainer",
     "//pkg/sql/colcontainer:colcontainer_test",
@@ -2357,6 +2358,7 @@ GET_X_DATA_TARGETS = [
     "//pkg/sql/catalog/systemschema_test:get_x_data",
     "//pkg/sql/catalog/tabledesc:get_x_data",
     "//pkg/sql/catalog/typedesc:get_x_data",
+    "//pkg/sql/clustermode:get_x_data",
     "//pkg/sql/clusterunique:get_x_data",
     "//pkg/sql/colcontainer:get_x_data",
     "//pkg/sql/colconv:get_x_data",
