+++ pkg/sql/BUILD.bazel
@@ -324,6 +324,7 @@ go_library(
         "//pkg/sql/catalog/systemschema",
         "//pkg/sql/catalog/tabledesc",
         "//pkg/sql/catalog/typedesc",
+        "//pkg/sql/clustermode",
         "//pkg/sql/clusterunique",
         "//pkg/sql/colexec",
         "//pkg/sql/colflow",
@@ -665,6 +666,7 @@ go_test(
         "//pkg/sql/catalog/systemschema",
         "//pkg/sql/catalog/tabledesc",
         "//pkg/sql/catalog/typedesc",
+        "//pkg/sql/clustermode",
         "//pkg/sql/clusterunique",
         "//pkg/sql/distsql",
         "//pkg/sql/execinfra",
