+++ pkg/sql/clustermode/BUILD.bazel
@@ -1,0 +1,15 @@
+load("//build/bazelutil/unused_checker:unused.bzl", "get_x_data")
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "clustermode",
+    srcs = ["cluster_mode.go"],
+    importpath = "github.com/cockroachdb/cockroach/pkg/sql/clustermode",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/settings",
+        "//pkg/sql/sessiondatapb",
+    ],
+)
+
+get_x_data(name = "get_x_data")
